requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
aiohttp>=3.9.1
matplotlib>=3.8.2

# Discord Bot Dependencies
discord.py>=2.0.0
python-dotenv
aiohttp

# Data Processing
matplotlib

# Configuration 
pyyaml  # For YAML config

# Optional: Better performance
# uvloop  # Unix only, faster async event loop